requirements.txt
pyrtlsdr
numpy
scipy
matplotlib
scikit-image
tqdm
asyncio